Move constants to separate module
Move constants to separate module so they can be used by other modules
too.
